Text enhancement: a truely worked codepage auto-detector
Previously, I've implemented a codepage auto-detector. But it did not
work properly so it was removed from the code base, and a user-configured
codepage selector was used.
Since now other languages are handled through the language-file mechanism,
we just have to detect the two variants of Chinese language now. Now,
by fixing the algorithm used, this function works perfectly for detecting
Simplified Chinese and Traditional Chinese.
As a result, the user-configured codepage selector is now useless and will
be removed in a future commit along with some other enhancements.