bugfix: optimize the player's use of attack magic.
Note: This is just a HACK ! ! ! ! ! ! 
In the original game, casting magic was an action with a layer concept, now it has the same effect as the original.
Layer calculations are performed for the image using Y coordinates and layer offsets.
Lack of experimental universality.  In the absence of testing, this adjustment is not necessarily correct.