bugfix: reconstructed the drawing scheme for the magical sprites of attack type magic.
Reconstructed the drawing scheme for the magical sprites of attack type magic, giving it a layer concept. It is no longer directly covering the screen, but calculating the number of layers. And sort enemies, players, and magic based on layer level. I have compared the playback effects of all the magics sprites in PAL.EXE and there are basically no problems.