fixed Multiple bugs.
fixbug: when the number of items you want to wear is 1 and the number of items you are wearing is also 1, replace them directly, instead of removing items and adding them at the end of the item menu.

fixbug: in the original DOS version, the text should be drawn with 3 layers shading, and the win95 version is a layer of shading.

fixbug: the font color of the description text for props and magic in sdlpal is different from that in the original Win95 version.