fixbug: The shortcut key "R" automatically selects the wrong enemy target.
when selecting player actions, the actual target does not exist and should be searched forward to attack the target.

when a team member performs an action, the actual target does not exist and should be searched backwards to attack the target.

When the shortcut key "R" repeats the previous round action (the shortcut key "R" covers the above two actions), the actual target does not exist, and the attack target should be searched forward and then backward again.